requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
Django==5.2.6
django-allauth==65.11.2
idna==3.10
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1